Fix test failures for SDwire preparation

+++ scripts/sd-wire.robot
@@ -16,7 +16,9 @@ Suite Teardown      Run Keyword    Log Out And Close Connection
 SDWire_001 SDWire recognition
     [Documentation]    This test aims to verify that the connected SDWire is
     ...    recognizable by the Test Server.
-    ${output}=    SDWire Diagnosis
+    ${output}    ${configed}=    SDWire Diagnosis
+    Set Suite Variable    ${configed}    ${configed}
+    Pass Execution If    ${configed}    Configured SD Wire detected. Skipping the IDs detection.
     ${vendor_id}    ${product_id}    ${serial_id}=    SDWire Identification    ${output}
     Set Suite Variable    ${vendor_id}    ${vendor_id}
     Set Suite Variable    ${product_id}    ${product_id}
@@ -26,11 +28,13 @@ SDWire_002 SDWire configuration and reading
     [Documentation]    This test aims to verify that the configuration
     ...    procedure for the SDWire works correctly and, after the
     ...    procedure, the test device is readable.
-    Configure SDWire    ${serial_id}    ${vendor_id}    ${product_id}
-    SSHLibrary.Write    reboot
-    Sleep    5s
-    Log Out And Close Connection
-    Setup SSH Connection
+    IF    not ${configed}
+        Configure SDWire    ${serial_id}    ${vendor_id}    ${product_id}
+        SSHLibrary.Write    reboot
+        Sleep    5s
+        Log Out And Close Connection
+        Setup SSH Connection
+    END
     Check SDWire Configuration
 
 SDWire_003 SDWire connecting to the Test Server
@@ -53,8 +57,7 @@ SDWire_005 SDWire connecting to the Device Under Test
 SDWire_006 OS booting form card mounted in the SDWire
     [Documentation]    This test aims to verify that the DUT boots properly
     ...    after flashing SD Card by using SDWire.
-    Change Relay State
-    Wait For Login Prompt In OS
+    Power DUT On And Wait For Login Prompt In OS
     Log Out And Close Connection
     Setup SSH Connection
-    Change Relay State
+    Relay Off
